ffmpeg: optionally allow minimal ffmpeg binary in image
When testing ffmpeg / hardware decoders it's very convenient
if one can do that with ffmpeg instead of only kodi. Using the
same ffmpeg patches and options as used by kodi also reduces the
risk of testing a different version than what's used by LE/kodi.

As the ffmpeg binary is way too minimal for normal use and only
intended for internal testing this feature is disabled by default
and has to be explicitly requested by building with FFMPEG_TESTING=yes

In addition to ffmpeg and ffprobe binaries this enables the null muxer
and wrapped_avframe encoder so one can test with "-f null".

On RPi it also enables vout_drm (which is only included in jc's
tree / RPi ffmpeg patches) so "-f vout_drm" can be used, too.

Signed-off-by: Matthias Reichl <hias@horus.com>