diffutils: update to 3.11
ann: https://lists.gnu.org/archive/html/info-gnu/2025-02/msg00003.html

does not rebuild with autotools 2.72 (as 3.11 was built with a newer version)
so patched Makefile.am causes the build failure. build is successful in GHA
whereas was previously failing on jenkins - thus the patch for no doc/man

removing the patch and going with configure instead of autotools reduces the
build tine.